v2.8 cotents
Switched to double space in place of tabs for indentation. Seems to be easier on the eyes, not to mention a more effective use of space. In the web community this also happens to be a more predominant style of indentation, especially among the big boys of IT world — Google HTML, CSS or XML style guides, for example.